requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.23.0
scikit-learn==1.5.1
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
google-generativeai==0.7.2